pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = ""
authors = ["rodrigo.guedes <gedsao77@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
django-admin-interface = "^0.22.1"
django-ordered-model = "^3.6"
python-decouple = "^3.6"
pytest-django = "^4.5.2"
pytest = "^7.1.3"
dj-database-url = "^1.0.0"
django-extensions = "^3.2.1"
gunicorn = "^20.1.0"
model-mommy = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
